Terminado a integração com a api da page teams

+++ src/components/PoPup/index.js
@@ -3,6 +3,7 @@ import iconDelete from "../../assets/poPup/iconDelete.svg"
 import { Container } from "./styles"
 
 import { api } from "../../services/api";
+import Swal from "sweetalert2";
 
 function PoPup({ display, setProps, id }) {
 
@@ -10,18 +11,50 @@ function PoPup({ display, setProps, id }) {
     const handleDelete = async (e) => {
         e.preventDefault();
 
-        try {
-            await api.delete(`/disciplines/${id}`);
-            setProps(true)
+        if (await confirmationToDelete()) {
+            try {
+                await api.delete(`/disciplines/${id}`);
+                setProps(true)
 
-        } catch (error) {
+            } catch (error) {
 
-            if (error.response === undefined) {
-                // httpError503()
-            } else {
+                if (error.response === undefined) {
+                    // httpError503()
+                } else {
 
+                }
             }
         }
+
+    }
+
+
+
+    const confirmationToDelete = async () => {
+        return await Swal.fire({
+            title: 'Tem certeza?',
+            text: "Você não poderá reverter isso!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#28a745',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sim, exclua!',
+            cancelButtonText: 'Não, exclua!'
+
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Excluído',
+                    'Disciplina excluído, com sucesso.',
+                    'success'
+                )
+                return true;
+
+            } else {
+                return false;
+            }
+
+        })
 
     }
